refactor(shared/components): heading, text를 typography로 통합
