--- packages/shared/components/src/Heading/index.ts
@@ -1,2 +1,0 @@
-/* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
-export { Heading, type HeadingProps } from './heading';
