+++ packages/shared/components/src/typography/index.ts
@@ -1,2 +1,4 @@
 /* eslint-disable @stylistic/padding-line-between-statements, react-refresh/only-export-components */
+export { Heading, type HeadingProps } from './heading';
+
 export { Text, type TextProps } from './text';
